Move everything out of NENV and destroy it
