+++ ironflow/ironflow/boxes/node_interface/input_widgets.py
@@ -10,7 +10,7 @@ from __future__ import annotations
 import ipywidgets as widgets
 
 from ironflow.ironflow.gui import GUI
-from ironflow.NENV import Node
+from ironflow.main.node import Node
 
 
 class SliderControl:
