+++ ironflow/nodes/std/special_nodes.py
@@ -1,7 +1,8 @@
 import code
 from contextlib import redirect_stdout, redirect_stderr
 
-from ironflow.NENV import import_widgets, NodeInputBP, NodeOutputBP, dtypes
+from ironflow.main import NodeInputBP, NodeOutputBP, dtypes
+from ironflow.main.utils import import_widgets
 from ironflow.main.node import Node
 from ironflow.ironflow.boxes.node_interface.input_widgets import SliderControl
